@@ -10,8 +10,15 @@ class Purges(Fuzzy.Cog):
         """Deletes the most recent messages in a channel.
         `amount` is the number of messages to delete."""
         amount += 1
-        deleted = await ctx.channel.purge(limit=amount, bulk=True,)
+        deleted = await ctx.channel.purge(
+            limit=amount,
+            bulk=True,
+        )
         await ctx.reply(
             f"Purged {len(deleted) - 1} messages from {ctx.channel.mention}.",
             delete_after=5,
         )
+
+
+async def setup(bot):
+    await bot.add_cog(Purges(bot))
